% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_output}
\alias{capture_output}
\title{Function to suppress output from cat() in getting spaMM summary
Borrowed / copied from purrr package, (don't want all the purrr dependencies)}
\usage{
capture_output(code)
}
\description{
Function to suppress output from cat() in getting spaMM summary
Borrowed / copied from purrr package, (don't want all the purrr dependencies)
}
