% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_de_bygene.R
\name{smi_de}
\alias{smi_de}
\title{Run DE analysis models on SMI data.}
\usage{
smi_de(
  assay_matrix,
  metadata,
  formula,
  neighborhood_counts = NULL,
  pre_de_obj = NULL,
  groupVar = NULL,
  groupVar_levels = NULL,
  nCores = 1,
  family = "nbinom2",
  targets = NULL,
  cellid_colname = "cell_ID",
  spatial_model = NULL,
  neighbor_expr_overlap_weight_colname = NULL,
  neighbor_expr_overlap_agg = c("sum", "mean"),
  neighbor_expr_cell_type_metadata_colname = "cell_type",
  neighbor_expr_totalcount_normalize = TRUE,
  neighbor_expr_totalcount_scalefactor = NULL,
  ...
)
}
\arguments{
\item{assay_matrix}{counts or normalized assay matrix.
Raw counts recommended for count distributions s.a. negative binomial (i.e., family=nbinom2).
Normalized is recommended for gaussian models.}

\item{metadata}{data.table or data.frame of meta.data associated with assay.
Must contain groupVar column and any covariates in the \code{formula} argument used for DE.
If cell id is not included as a column in metadata, should correspond to rownames in the metadata data.frame object.}

\item{formula}{right hand side of formula used in DE, possibly containing random effects.
Must contain groupVar fixed effect.
It's recommended that raw count models with a log-link (i.e., negative binomial,poisson)
, use a corresponding offset for library size in formula
(i.e., ~de_variable + offset(log(totalcounts))), where totalcounts is a column in the metadata.}

\item{neighborhood_counts}{Deprecated argument (will be removed in the future); optional object created by smiDE::measure_neighbor_expression_by_celltype, which contains lists of
the expression counts in neighboring cells by cell type compared to a cell type of reference
(see \code{help(measure_neighbor_expression_by_celltype)}).
May be used to control for confounding or 'bleed-over' in DE model (see examples)}

\item{pre_de_obj}{optional list object created by smiDE::pre_de, which contains a table of cell_adjacencies \code{cell_adjacency_dt} used for calculating neighbor-cell expression,
and, if created with smiDE::pre_de(adjacencies_only=FALSE), an \code{nblist} object with pre-computed neighbor-cell expressions.
(see \code{help(pre_de)} or \code{help(measure_neighbor_expression_by_celltype)} for details).
May be used to control for confounding or 'bleed-over' in DE model (see examples)}

\item{groupVar}{optional categorical variable for which you may specify the ordering of factor levels, this column must exist in meta.data slot}

\item{groupVar_levels}{optional factor ordering for groupVar.}

\item{nCores}{= 1, number of cores to use, set to 1 if running in serial mode}

\item{family}{a character string naming a family function,
See \code{family} for a generic discussion of families.
Default is nbinom2 (with log link), specifying Negative binomial distribution: quadratic parameterization (Hardin & Hilbe 2007). V=mu*(1+mu/phi) = mu+mu^2/phi.}

\item{targets}{optional vector of targets to run DE for.  If not supplied, run models for all targets in assay.}

\item{cellid_colname}{column name in metadata corresponding to cell id.}

\item{spatial_model}{optional list for fitting models with spatially correlated random effects.
This list must have "name" (indicating the type of spatial random effect model), as a minimum requirement.
See documentation and examples for details.
\link{spatial_model}}

\item{neighbor_expr_overlap_weight_colname}{optional argument used in the case of provided \code{pre_de_obj} created with \code{cell_adjacencies_only=TRUE}.
Used for neighbor expression control variable, default is NULL (unweighted), but can also be specified by a weight column in the \code{cell_adjacency_dt}
(a column 'weight' is pre-computed in the \code{cell_adjacency_dt} as \code{1/distance} between neighboring cells).
Cells with higher weights (closer to the modeled cells) can be given more
weight in computing the neighbor expression of a given gene.  Coincides with assumption that closer cells are more likely to overlap/cause segmentation errors with the modeled cells.}

\item{neighbor_expr_overlap_agg}{whether to use the mean "sum" (sum total) or average "mean" of neighbor cell expression as the control variable. Default is "sum".}

\item{neighbor_expr_cell_type_metadata_colname}{column to use for determining neighbor cell expression by cell type.  Represented in pre_de_obj$cell_adjacency_dt}

\item{neighbor_expr_totalcount_normalize}{Defaults to TRUE, neighbor expression covariate is normalized by the library size or 'totalcounts' across all genes for a given cell.
y_norm{cell,gene} = y_counts{cell,gene} * mean(totalcounts{allcells}) / totalcounts{cell}}

\item{neighbor_expr_totalcount_scalefactor}{If NULL, and neighbor_expr_totalcount_normalize is set to TRUE, this will calculate the cell-specific scaling factors for normalization directly from \code{assay_matrix}.}

\item{...}{further arguments to be passed to model fitting function.
See example below.}
}
\value{
data.table of DE results, with one row per target, marginal means and their SE's of the DE groups
, estimated fold change and p-values for non-zero difference (identity link) or ratio \eqn{\neq 1} (log link) between groups.
}
\description{
Run DE analysis models on SMI data.
}
\examples{

library(Seurat)
library(data.table); setDTthreads(1)
datadir<-system.file("extdata", package="smiDE")
sem <- readRDS(paste0(datadir, "/small_nsclc.rds"))
metainfo <- data.table(sem@meta.data)
totalcount_scalefactors <- mean(metainfo[["totalcounts"]]) / metainfo[["totalcounts"]]
names(totalcount_scalefactors) <- colnames(sem)
sem <- Seurat::SetAssayData(sem
                           ,"data"
                           ,sem[["RNA"]]@counts \%*\% Matrix::Diagonal(x=totalcount_scalefactors, names=colnames(sem))
                           )

pre_de_obj <- 
pre_de(metadata = metainfo
       ,cell_type_metadata_colname = "cell_type"
       ,split_neighbors_by_colname = "tissue"
       ,mm_radius = 0.05
       ,sdimx_colname = "sdimx"
       ,sdimy_colname = "sdimy"
       ,verbose=TRUE
)

fibroblast_and_macrophage_cells <- metainfo[cell_type \%in\% c("fibroblast", "macrophage"),cell_ID]
de_results <-       
  smi_de(assay_matrix = sem[["RNA"]]@counts
         ,metadata = metainfo[cell_ID \%in\% fibroblast_and_macrophage_cells]
         ,formula = ~RankNorm(otherct_expr) + niche + tissue + offset(log(totalcounts)) 
         ,family="nbinom2"
         ,targets=rownames(sem)[1:5]
         ,pre_de_obj = pre_de_obj
         ,neighbor_expr_cell_type_metadata_colname = "cell_type"
         ,neighbor_expr_overlap_weight_colname = NULL
         ,neighbor_expr_overlap_agg ="sum"
         ,neighbor_expr_totalcount_normalize = TRUE
         ,neighbor_expr_totalcount_scalefactor = totalcount_scalefactors
         )  
  ) 

results(de_results, "pairwise", variable="niche", targets = rownames(sem)[1:2])
results(de_results, "one.vs.rest", variable="tissue", targets = rownames(sem)[1:2])

## fit a mixed model with all cells, cell type as a fixed effect covariate, 
## tissue/sample id as a random effect covariate

de_results <-       
  smi_de(assay_matrix = sem[["RNA"]]@counts
         ,metadata = metainfo
         ,formula = ~RankNorm(otherct_expr) + cell_type + (1 | tissue) + offset(log(totalcounts)) 
         ,family="nbinom2"
         ,targets=rownames(sem)[1:5]
         ,pre_de_obj = pre_de_obj
         ,neighbor_expr_cell_type_metadata_colname = "cell_type"
         ,neighbor_expr_overlap_weight_colname = NULL
         ,neighbor_expr_overlap_agg ="sum"
         ,neighbor_expr_totalcount_normalize = TRUE
         ,neighbor_expr_totalcount_scalefactor = totalcount_scalefactors
         )  
  ) 
results(de_results, "pairwise", variable="cell_type", targets = rownames(sem)[1:2])
results(de_results, "one.vs.rest", variable="cell_type", targets = rownames(sem)[1:2])
  
  
}
